parser/requirements.txt
attrs==23.1.0
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
dnspython==2.4.2
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
packaging==23.2
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pymongo==4.6.1
pyOpenSSL==23.3.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
tldextract==5.1.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.1.0
w3lib==2.1.2
zope.interface==6.1
